+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 0.1.39 (TBD)
-- TODO
+- Fixed issue with texture ribbon being overwritten by M2 textures when viewing a WMO.
 
 0.1.38 (13-09-2021)
 - Implemented remote control protocol for advanced users.