+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.1.39 (TBD)
+0.1.39 (14-09-2021)
 - Added option to export M2 bone data into a relative *_bones.json file.
 - Added an `exportID` property to `EXPORT_` and `HOOK_EXPORT_COMPLETE` RCP messages.
 - M2 models exported with a skin will now be named uniquely based on the selected skin.